# Do not use indices LUT in Batched Curl