--- a/curl/bct/transform_amd64.s
+++ b/curl/bct/transform_amd64.s
@@ -2,75 +2,62 @@
 
 #include "textflag.h"
 
-// func transform(lto, hto, lfrom, hfrom *[729]uint, rounds uint)
+#define SBOX(la, ha, lb, hb) \
+    MOVQ lb, R14; \             // a := (lb ^ ha) & la
+    XORQ ha, R14; \
+    ANDQ la, R14; \
+    MOVQ hb, R15; \             // b := (hb ^ la) | a
+    XORQ la, R15; \
+    ORQ R14, R15
+
+// func transform(lto, hto, lfrom, hfrom *[729]uint)
 TEXT ·transform(SB),NOSPLIT,$0
     MOVQ lto+0(FP), AX
     MOVQ hto+8(FP), BX
     MOVQ lfrom+16(FP), CX
     MOVQ hfrom+24(FP), DX
-    MOVQ rounds+32(FP), SI      // r = rounds
-    TESTQ SI, SI                // if r == 0 goto DONE
-    JZ DONE
-
-LOOP:
-    MOVQ $·Indices(SB), R8
-    MOVQ (R8), R8
-    MOVQ (CX)(R8*8), R14        // ltmp := lfrom[Indices[0]]
-    MOVQ (DX)(R8*8), R15        // htmp := hfrom[Indices[0]]
-
-    XORQ DI, DI                 // i := 0
+    MOVQ $81, SI                // r := curl.NumRounds
 
 ROUND:
-    MOVQ $·Indices(SB), R12
-    MOVQ 8(R12)(DI*8), R9       // t1 := Indices[i+1]
-    MOVQ 16(R12)(DI*8), R10     // t2 := Indices[i+2]
-    MOVQ 24(R12)(DI*8), R11     // t3 := Indices[i+3]
+    MOVQ (CX), R10              // l0 := lfrom[0]
+    MOVQ (DX), R11              // h0 := hfrom[0]
+    MOVQ 2912(CX), R12          // l1 := lfrom[364]
+    MOVQ 2912(DX), R13          // h1 := hfrom[364]
 
-    MOVQ R14, R12               // l0 := ltmp
-    MOVQ R15, R13               // h0 := htmp
-    MOVQ (CX)(R9*8), R14        // l1 := lfrom[t1]
-    MOVQ (DX)(R9*8), R15        // h1 := hfrom[t1]
-    MOVQ (CX)(R10*8), R8
-    MOVQ (DX)(R10*8), R9
-    MOVQ (CX)(R11*8), R10
-    MOVQ (DX)(R11*8), R11
+    SBOX(R10, R11, R12, R13)    // a, b := sBox(l0, h0, l1, h1)
+    MOVQ R15, (BX)              // hto[0] = b
+    NOTQ R14                    // lto[0] = ^a
+    MOVQ R14, (AX)
 
-    XORQ R14, R13               // a0 := (l1 ^ h0) & l0
-    ANDQ R12, R13
-    XORQ R15, R12               // b0 := (h1 ^ l0) | a0
-    ORQ R13, R12
-    MOVQ R12, (BX)(DI*8)        // hto[i] = b0
-    NOTQ R13                    // lto[i] = ^a0
-    MOVQ R13, (AX)(DI*8)
+    MOVQ $364, R8               // t := 364
+    MOVQ $1, DI                 // i := 1
 
-    XORQ R8, R15
-    ANDQ R14, R15
-    XORQ R9, R14
-    ORQ R15, R14
-    MOVQ R14, 8(BX)(DI*8)
-    NOTQ R15
-    MOVQ R15, 8(AX)(DI*8)
+LOOP:
+    MOVQ 2912(CX)(R8*8), R10    // l0 = lfrom[t+364]
+    MOVQ 2912(DX)(R8*8), R11    // h0 = hfrom[t+364]
 
-    XORQ R10, R9
-    ANDQ R8, R9
-    XORQ R11, R8
-    ORQ R9, R8
-    MOVQ R8, 16(BX)(DI*8)
-    NOTQ R9
-    MOVQ R9, 16(AX)(DI*8)
+    SBOX(R12, R13, R10, R11)    // a, b = sBox(l1, h1, l0, h0)
+    MOVQ R15, (BX)(DI*8)        // hto[i] = b
+    NOTQ R14                    // lto[i] = ^a
+    MOVQ R14, (AX)(DI*8)
 
-    MOVQ R10, R14               // ltmp = lfrom[t3]
-    MOVQ R11, R15               // htmp = hfrom[t3]
+    MOVQ -8(CX)(R8*8), R12       // l1 := lfrom[t-1]
+    MOVQ -8(DX)(R8*8), R13       // h1 := hfrom[t-1]
 
-    ADDQ $3, DI                 // i += 3
-    CMPQ DI, $727               // if i < 727 goto ROUND
-    JL ROUND
+    SBOX(R10, R11, R12, R13)    // a, b = sBox(l0, h0, l1, h1)
+    MOVQ R15, 8(BX)(DI*8)       // hto[i+1] = b
+    NOTQ R14                    // lto[i+1] = ^a
+    MOVQ R14, 8(AX)(DI*8)
+
+    DECQ R8                     // t--
+    ADDQ $2, DI                 // i += 2
+    CMPQ DI, $728               // if i < 728 goto ROUND
+    JL LOOP
 
     XCHGQ AX, CX                // lfrom, lto = lto, lfrom
     XCHGQ BX, DX                // hfrom, hto = hto, hfrom
 
     DECQ SI                     // r--
-    JNZ LOOP                    // if r != 0 goto LOOP
+    JNZ ROUND                   // if r != 0 goto LOOP
 
-DONE:
     RET