--- a/curl/bct/transform_generic.go
+++ b/curl/bct/transform_generic.go
@@ -2,10 +2,6 @@
 
 package bct
 
-import (
-	"github.com/iotaledger/iota.go/curl"
-)
-
-func transform(lto, hto, lfrom, hfrom *[curl.StateSize]uint, rounds uint) {
-	transformGeneric(lto, hto, lfrom, hfrom, rounds)
+func transform(lto, hto, lfrom, hfrom *[729]uint) {
+	transformGeneric(lto, hto, lfrom, hfrom)
 }