# Description of change

This PR gets rid of the LUT `curl.Indices` in the batched Curl transpose and computes it value on the fly. In order to avoid costly modulo operations or conditionals it unrolls an even number of iterations in the state loop. That way, the indices can be computed by alternately add 364 or subtract 365.

The leads to a 10% speedup on `amd64` compared to the current master:
```
name                old time/op  new time/op  delta
CurlTransaction     3.18ms ±10%  2.86ms ± 5%  -10.29%
CurlHash             115µs ± 4%   106µs ± 6%   -7.85%
```
By default, `amd64` uses the Assembly version; the speedup of the pure Go implementation is even 20%:
```
name                old time/op  new time/op  delta
CurlTransaction     3.87ms ± 4%  3.05ms ± 5%  -21.25%
CurlHash             140µs ± 8%   115µs ± 7%  -18.31%
```
## Type of change

- Enhancement (a non-breaking change which adds functionality)

## How the change has been tested

The existing test suite has been extended to test the batched Curl also against the regular Curl golden test vectors as single input.

## Change checklist

- [x] My code follows the contribution guidelines for this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have added tests using ginkgo that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes