@Wollac 2021-03-10
With this PR the performance improvements of the `amd64` Assembly version have decreased further. It has gone done to less than 8% when comparing Assembly to pure pure Go:
```
name                old time/op  new time/op  delta
CurlTransaction     3.05ms ± 5%  2.86ms ± 5%  -6.28%
CurlHash             115µs ± 7%   106µs ± 6%  -7.65%
```
This introduces the question what speed-up still justifies having very hard to maintain and platform dependent assembly code for `amd64`. Or should it maybe removed, as this is also the only package in `iota.go` that includes Assembly code.