@luca-moser 2021-03-30
If the new Go impl. is faster than the assembly code, then there is no need to longer have it in the lib.