@Wollac 2021-04-01
> If the new Go impl. is faster than the assembly code, then there is no need to longer have it in the lib.

The assembly version is still 4% - 5% faster than the pure go version. The following compares `master` against the pure go verstion from `Wollac:curl-indices`:
```
name                old time/op  new time/op  delta
CurlTransaction-48  3.18ms ±10%  3.05ms ± 5%  -4.28%  (p=0.007 n=16+16)
CurlHash-48          115µs ± 4%   115µs ± 7%    ~     (p=0.654 n=15+16)
```
However, this is really not that much and the pure go version from `Wollac:curl-indices` is still 6%-7% faster than the previous assembly version. That's exactly why I have been wondering whether keeping and maintaining an `amd64` assembly version still makes sense for a mere 4% speedup...