@luca-moser 2021-04-06
Okay, now I got. Given the fact that the implementation shouldn't really change in the future, I don't think it is really necessary to remove it, even though you're probably the only person who understands the assembly version. I'd leave it in.